# openapi.yml
openapi: 3.0.0
info:
  title: MiraiCore API
  version: 1.0.0
servers:
  - url: http://{host}:{port}/api
    variables:
      host:
        default: localhost
      port:
        default: 80
paths:
  /ping:
    get:
      summary: Ping the server
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  Status:
                    type: string
                  message:
                    type: string
  /signin:
    post:
      summary: Sign in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uid:
                  type: string
                email:
                  type: string
                pass:
                  type: string
                photoUrl:
                  type: string
      responses:
        '200':
          description: Successful sign in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid request payload
        '415':
          description: Unsupported Media Type
        '500':
          description: Internal Server Error
  /GoSchool:
    get:
      summary: Mark attendance for a user
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful attendance
          content:
            application/json:
              schema:
                type: object
                properties:
                  Status:
                    type: string
                  message:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /LessonDetails:
    get:
      summary: Get lesson details
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
        - in: query
          name: ClassID
          schema:
            type: string
        - in: query
          name: StartDate
          required: true
          schema:
            type: string
        - in: query
          name: EndDate
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  Monday:
                    type: array
                    items:
                      type: object
                      properties:
                        ClassID:
                          type: string
                        DayOfTheWeek:
                          type: string
                        LessonNumber:
                          type: integer
                        Lesson:
                          type: string
                        Room:
                          type: string
                        Teacher:
                          type: string
                        StartTime:
                          type: string
                        EndTime:
                          type: string
                  Tuesday:
                    type: array
                    items:
                      type: object
                      properties:
                        ClassID:
                          type: string
                        DayOfTheWeek:
                          type: string
                        LessonNumber:
                          type: integer
                        Lesson:
                          type: string
                        Room:
                          type: string
                        Teacher:
                          type: string
                        StartTime:
                          type: string
                        EndTime:
                          type: string
        '400':
          description: Missing required parameters
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
        '404':
          description: Not Found